refactor: reshape types.

+++ projects/games/battle/include/constants.h
@@ -1,8 +1,6 @@
 #ifndef CONSTANTS_H
 #define CONSTANTS_H
 
-#include <stdint.h>
-
-extern const uint16_t ATTACK_RANGE;
+extern const int ATTACK_RANGE;
 
 #endif // CONSTANTS_H
