+++ projects/games/battle/src/entity.c
@@ -2,7 +2,7 @@
 #include "entity.h"
 #include "random.h"
 
-static int16_t clamp_attack_value(const int16_t value) {
+static int clamp_attack_value(const int value) {
     if (value <= 0) {
         return 1;
     } else {
@@ -14,8 +14,8 @@ bool entity_is_alive(const entity *e) {
     return e->current_health > 0;
 }
 
-uint16_t entity_get_attack_value(const entity *e, const entity_attack_value_type type) {
-    int16_t value = 0;
+int entity_get_attack_value(const entity *e, const entity_attack_value_type type) {
+    int value = 0;
 
     switch (type) {
         case ENTITY_ATTACK_VALUE_MINIMUM:
@@ -41,7 +41,7 @@ uint16_t entity_get_attack_value(const entity *e, const entity_attack_value_type
             break;
     }
 
-    return (uint16_t)clamp_attack_value(value);
+    return (int)clamp_attack_value(value);
 }
 
 bool entity_is_critical_hit(const entity *e) {
@@ -49,8 +49,8 @@ bool entity_is_critical_hit(const entity *e) {
 }
 
 // Returns the actual damage taken after defense calculations
-uint16_t entity_take_damage(entity *e, const uint16_t damage) {
-    int16_t value = (int16_t)damage;
+int entity_take_damage(entity *e, const int damage) {
+    int value = (int)damage;
 
     // If the entity is defending, the damage is halved.
     if (e->is_defending) {
@@ -68,8 +68,8 @@ uint16_t entity_take_damage(entity *e, const uint16_t damage) {
     if (value >= e->current_health) {
         e->current_health = 0;
     } else {
-        e->current_health -= (uint16_t)value;
+        e->current_health -= (int)value;
     }
 
-    return (uint16_t)value;
+    return (int)value;
 }
